components/codeupipe-google-ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codeupipe-google-ai"
version = "0.1.0"
description = "Google AI (Gemini) — multimodal generation, embeddings, and vision"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "codeuchain"}]
dependencies = ["codeupipe>=0.8.0", "google-generativeai>=0.4.0"]

[project.urls]
Homepage = "https://github.com/codeuchain/codeupipe-marketplace"
Repository = "https://github.com/codeuchain/codeupipe-marketplace"
